Refactor recombination funtions into their own files
