+++ CMakeLists.txt
@@ -32,6 +32,7 @@ add_library(recombo_core STATIC
         src/legacyBfacf.cpp
         src/bfacfMove.cpp
         src/edgePool.cpp
+        src/clkRecombination.cpp
         src/mmc_config.cpp
         src/mmchain.cpp
         src/newsud.cpp
